+++ src/components/Feedback/Dialog/DialogCloseIcon/DialogCloseIcon.tsx
@@ -1,5 +1,5 @@
 import { IconButton } from "@mui/material";
-import CloseIcon from "@/components/Icon/CloseIcon";
+import CloseIcon from "@/components/DataDisplay/Icons/CloseIcon";
 
 export interface DialogCloseProps {
   onClick: () => void;