+++ src/components/Navigation/NavigationMenu/BurgerAppBar/BurgerAppBar.tsx
@@ -1,17 +1,17 @@
 import { AppBar, GlobalStyles, IconButton, Stack, Toolbar, useTheme } from "@mui/material";
-import { useContext } from "react";
-import { NavigationMenuContext, SecondaryMenu } from "@/components/Navigation/NavigationMenu";
-import MenuIcon from "@/components/Navigation/NavigationMenu/MenuIcon";
-import SecondaryMenuButton from "@/components/Navigation/NavigationMenu/SecondaryMenuButton";
+import { cloneElement, useContext } from "react";
+import MenuIcon from "@/components/DataDisplay/Icons/MenuIcon";
+import { NavigationMenuContext } from "@/components/Navigation/NavigationMenu";
 
-interface TabletNavBarProps {
-  secondaryMenu?: SecondaryMenu;
-}
-
-const TabletNavBar = ({ ...props }: TabletNavBarProps) => {
-  const { openDrawerMenu, secondaryMenu = props.secondaryMenu } = useContext(NavigationMenuContext);
+const BurgerAppBar = () => {
+  const { openDrawerMenu, AppBar: AppBarComponent } = useContext(NavigationMenuContext);
   const { palette, size } = useTheme();
   const background = palette.mode === "dark" ? palette.background.default : palette.primary.black;
+  const AppBarWithTrigger = AppBarComponent ? cloneElement(AppBarComponent, { onClickBurger: openDrawerMenu }) : AppBarComponent;
+
+  if (AppBarComponent) {
+    return AppBarWithTrigger;
+  }
 
   return (
     <AppBar
@@ -27,9 +27,6 @@ const TabletNavBar = ({ ...props }: TabletNavBarProps) => {
         <IconButton onClick={openDrawerMenu} edge="end" color="inherit" aria-label="menu">
           <MenuIcon />
         </IconButton>
-        <Stack direction="row" alignItems="center" spacing={1}>
-          {secondaryMenu && <SecondaryMenuButton variant="icon" />}
-        </Stack>
       </Toolbar>
       <GlobalStyles
         styles={{
@@ -42,4 +39,4 @@ const TabletNavBar = ({ ...props }: TabletNavBarProps) => {
   );
 };
 
-export default TabletNavBar;
+export default BurgerAppBar;