+++ src/components/Navigation/NavigationMenu/BurgerAppBar/index.ts
@@ -1,0 +1,4 @@
+import BurgerAppBar from "./BurgerAppBar";
+
+export default BurgerAppBar;
+export * from "./BurgerAppBar";