+++ src/components/Navigation/NavigationMenu/MenuIcon/index.ts
@@ -1,3 +1,0 @@
-import MenuIcon from "./MenuIcon";
-
-export default MenuIcon;