+++ src/components/Navigation/NavigationMenu/MobileNavBar/index.ts
@@ -1,4 +1,0 @@
-import MobileNavBar from "./MobileNavBar";
-
-export default MobileNavBar;
-export * from "./MobileNavBar";