+++ src/components/Navigation/NavigationMenu/NavigationMenu.tsx
@@ -1,13 +1,12 @@
 import { SwipeableDrawer, useMediaQuery, useTheme } from "@mui/material";
-import { createContext, memo, ReactNode, useCallback, useContext, useMemo, useState } from "react";
-import MobileNavBar from "@/components/Navigation/NavigationMenu/MobileNavBar";
+import { createContext, memo, ReactElement, ReactNode, useCallback, useContext, useMemo, useState } from "react";
+import BurgerAppBar from "@/components/Navigation/NavigationMenu/BurgerAppBar";
 import SideBar from "@/components/Navigation/NavigationMenu/SideBar";
 import SideBarMenu from "@/components/Navigation/NavigationMenu/SideBarMenu";
-import TabletNavBar from "@/components/Navigation/NavigationMenu/TabletNavBar";
 
 export type ObjectNavigationItem = {
   url?: string;
-  label: string;
+  label?: string;
   count?: number;
   icon?: ReactNode;
   active?: boolean;
@@ -27,57 +26,11 @@ export interface NavLinkProps {
 
 export type NavigationItem = ObjectNavigationItem | ReactNode;
 
-export interface SecondaryMenu {
-  /**
-   * Menu id
-   */
-  id?: string;
-  /**
-   * Menu label
-   */
-  label?: string;
-  /**
-   * Menu sub label
-   */
-  subLabel?: string;
-  /**
-   * Menu start icon
-   */
-  startIcon?: ReactNode;
-  /**
-   * Menu icon
-   */
-  iconOpenMenu?: ReactNode;
-  /**
-   * Avatar props
-   */
-  avatar?: {
-    name?: string;
-    src?: string;
-  };
-  /**
-   * Menu items secondary menu
-   */
-  items: NavigationItem[];
-  /**
-   * Loading state
-   */
-  loading?: boolean;
-}
-
 export interface NavigationMenuProps {
   /**
    * Menu items
    */
   items?: NavigationItem[];
-  /**
-   * Menu items for mobile bottom navigation
-   */
-  itemsMobile?: NavigationItem[];
-  /**
-   * Menu items for secondary menu
-   */
-  secondaryMenu?: SecondaryMenu;
   /**
    * Override the default translations
    */
@@ -139,6 +92,17 @@ export interface NavigationMenuProps {
     hideNavBarOnRoutes?: string[];
   };
   /**
+   * Component to render the bottom link
+   */
+  bottomLink?: {
+    url?: string;
+    state?: any;
+    end?: boolean;
+    label?: ReactNode;
+    active?: boolean;
+    icon?: ReactNode;
+  };
+  /**
    * Component to router nav links.
    * This component is used to render the links in the main menu &  mobile bottom navigation
    * It should be a react-router-dom NavLink or a compatible component
@@ -147,10 +111,6 @@ export interface NavigationMenuProps {
    */
   NavLink?: (props: NavLinkProps) => ReactNode;
   /**
-   * Footer component only for desktop
-   */
-  Footer?: ReactNode;
-  /**
    * Logo component
    */
   Logo?: ReactNode;
@@ -158,6 +118,10 @@ export interface NavigationMenuProps {
    * Search component
    */
   Search?: ReactNode;
+  /**
+   * AppBar component for small screens
+   */
+  AppBar?: ReactElement;
 }
 
 const DEFAULT_CONTEXT_VALUE = {
@@ -173,34 +137,13 @@ type NavigationMenuContextValue = NavigationMenuProps & typeof DEFAULT_CONTEXT_V
 export const NavigationMenuContext = createContext<NavigationMenuContextValue>(DEFAULT_CONTEXT_VALUE);
 
 const NavigationMenuFactory = () => {
-  const { items, disableResponsive, isMobile, isTablet, itemsMobile, isDrawerOpen, closeDrawerMenu, sideBarWidth, openDrawerMenu } =
+  const { items, disableResponsive, isMobile, isTablet, isDrawerOpen, closeDrawerMenu, sideBarWidth, openDrawerMenu } =
     useContext(NavigationMenuContext);
 
-  if (isMobile && !disableResponsive) {
+  if ((isMobile || isTablet) && !disableResponsive) {
     return (
       <>
-        <MobileNavBar items={itemsMobile} />
-        <SwipeableDrawer
-          anchor="left"
-          open={isDrawerOpen}
-          onClose={closeDrawerMenu}
-          onOpen={openDrawerMenu}
-          PaperProps={{
-            sx: { width: "85%" },
-          }}
-        >
-          <SideBar width="100%">
-            <SideBarMenu items={items} />
-          </SideBar>
-        </SwipeableDrawer>
-      </>
-    );
-  }
-
-  if (isTablet && !disableResponsive) {
-    return (
-      <>
-        <TabletNavBar />
+        <BurgerAppBar />
         <SwipeableDrawer anchor="left" open={isDrawerOpen} onClose={closeDrawerMenu} onOpen={openDrawerMenu}>
           <SideBar>
             <SideBarMenu items={items} />
@@ -219,22 +162,21 @@ const NavigationMenuFactory = () => {
 
 const NavigationMenu = ({
   items,
-  itemsMobile,
-  secondaryMenu,
   enableSearchFocusShortcut,
   translations,
   disableResponsive,
   mobileOptions,
   hideSearchDesktop,
   NavLink,
-  Footer,
   Search,
   Logo,
+  AppBar,
+  bottomLink,
   sideBarWidth = 260,
 }: NavigationMenuProps) => {
-  const theme = useTheme();
-  const isMobile = useMediaQuery(theme.breakpoints.down("sm"));
-  const isTablet = useMediaQuery(theme.breakpoints.between("sm", "md"));
+  const { breakpoints } = useTheme();
+  const isMobile = useMediaQuery(breakpoints.down("sm"));
+  const isTablet = useMediaQuery(breakpoints.between("sm", "md"));
   const [isDrawerOpen, setIsDrawerOpen] = useState(DEFAULT_CONTEXT_VALUE.isDrawerOpen);
 
   const closeDrawerMenu = useCallback(() => {
@@ -247,22 +189,21 @@ const NavigationMenu = ({
 
   const value = useMemo(
     () => ({
+      AppBar,
+      bottomLink,
       closeDrawerMenu,
       disableResponsive,
       enableSearchFocusShortcut,
-      Footer,
       hideSearchDesktop,
       isDrawerOpen,
       isMobile,
       isTablet,
       items,
-      itemsMobile,
       Logo,
       mobileOptions,
       NavLink,
       openDrawerMenu,
       Search,
-      secondaryMenu,
       sideBarWidth,
       translations,
     }),
@@ -273,18 +214,17 @@ const NavigationMenu = ({
       isMobile,
       isTablet,
       items,
-      itemsMobile,
-      secondaryMenu,
       openDrawerMenu,
       sideBarWidth,
       translations,
-      Footer,
       Logo,
       NavLink,
       enableSearchFocusShortcut,
       mobileOptions,
       Search,
       hideSearchDesktop,
+      bottomLink,
+      AppBar,
     ],
   );