+++ src/components/Navigation/NavigationMenu/SecondaryMenuButton/index.ts
@@ -1,4 +1,0 @@
-import SecondaryMenuButton from "./SecondaryMenuButton";
-
-export default SecondaryMenuButton;
-export * from "./SecondaryMenuButton";