+++ src/components/Navigation/NavigationMenu/TabletNavBar/index.ts
@@ -1,4 +1,0 @@
-import TabletNavBar from "./TabletNavBar";
-
-export default TabletNavBar;
-export * from "./TabletNavBar";